requirements.txt
streamlit>=1.33.0
git+https://github.com/thomluther/anker-solix-api.git
aiohttp>=3.9.3
python-dotenv>=1.0.1
paho-mqtt>=1.6.1
plotly>=5.18.0
